fix(opencode): move frontmatter parser out of plugins/ directory (#596)
OpenCode's legacy plugin loader scans .opencode/plugins/ for modules and
treats every exported function as a plugin. ponytail-frontmatter.cjs
lived in that directory and exported parseCommandFile, which the loader
invoked with the plugin context object as its first argument, passing
{ client: ... } to fs.readFileSync. This threw "path must be a string
or a file descriptor" (#301, #596).

Move the parser one level up to .opencode/ponytail-frontmatter.cjs, outside
the scanned plugins/ directory. The plugins/ directory now contains only
ponytail.mjs with its single default export.
