+++ .opencode/ponytail-frontmatter.cjs
@@ -2,13 +2,11 @@
 
 // ponytail command-file frontmatter parser.
 //
-// Pulled out of ponytail.mjs so the plugin module's only top-level export is
-// the plugin function itself. OpenCode's legacy plugin loader (the one that
-// runs before v1 plugins are detected) treats every function exported from a
-// plugin module as a plugin; calling the frontmatter parser as one threw
-// "path must be a string or a file descriptor" because it got the plugin
-// context object as its first argument. Keeping the parser in its own module
-// leaves exactly one plugin-shaped export on ponytail.mjs.
+// Lives outside .opencode/plugins/ so OpenCode's legacy plugin loader
+// doesn't discover it. The legacy loader treats every exported function
+// from any file under plugins/ as a plugin; calling parseCommandFile with
+// the plugin context object threw "path must be a string or a file
+// descriptor" (#301, #596).
 
 function parseCommandFile(filePath) {
   const fs = require('fs');
